Fix for issue #1675. Added missing parent index calculation.
