+++ Tests/Linq/Linq/TakeSkipTests.cs
@@ -1,9 +1,10 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
 using LinqToDB;
-
+using LinqToDB.Mapping;
 using NUnit.Framework;
 
 namespace Tests.Linq
@@ -480,5 +481,69 @@ namespace Tests.Linq
 					);
 			}
 		}
+
+		public class Batch
+		{
+			[PrimaryKey]
+			public int Id { get; set; }
+			[Column]
+			public string Value { get; set; }
+
+			[Association(ThisKey = "Id", OtherKey = "BatchId", CanBeNull = false)]
+			public List<Confirmation> Confirmations { get; set; }
+		}
+
+		public class Confirmation
+		{
+			[Column]
+			public int BatchId { get; set; }
+			[Column]
+			public DateTime Date { get; set; }
+		}
+
+		[Test]
+		public void FirstOrDefaultInSubQuery([IncludeDataSources(TestProvName.AllSQLite)] string context)
+		{
+			using (var db = GetDataContext(context))
+			{
+				using (db.CreateLocalTable(new[]
+				{
+					new Batch { Id = 1, Value = "V1" },
+					new Batch { Id = 2, Value = "V2" },
+					new Batch { Id = 3, Value = "V3" }
+				}))
+				using (db.CreateLocalTable(new[]
+				{
+					new Confirmation { BatchId = 1, Date = DateTime.Parse("09 Apr 2019 14:30:00 GMT") },
+					new Confirmation { BatchId = 2, Date = DateTime.Parse("09 Apr 2019 14:30:20 GMT") },
+					new Confirmation { BatchId = 2, Date = DateTime.Parse("09 Apr 2019 14:30:25 GMT") },
+					new Confirmation { BatchId = 3, Date = DateTime.Parse("09 Apr 2019 14:30:35 GMT") },
+				}))
+				{
+				
+					var query = db.GetTable<Batch>()
+							.OrderByDescending(x => x.Id)
+							.Select(x => new
+							{
+								BatchId = x.Id,
+								CreationDate = x.Confirmations.FirstOrDefault().Date,
+								x.Value
+							})
+							.Take(2)
+							.OrderBy(x => x.BatchId);
+
+					var res = query.ToList();
+
+					Assert.That(res.Count,           Is.EqualTo(2));
+					Assert.That(res[0].BatchId,      Is.EqualTo(2));
+					Assert.That(res[0].Value,        Is.EqualTo("V2"));
+					Assert.That(res[1].BatchId,      Is.EqualTo(3));
+					Assert.That(res[1].Value,        Is.EqualTo("V3"));
+					Assert.That(res[0].CreationDate, Is.EqualTo(DateTime.Parse("09 Apr 2019 14:30:20 GMT")));
+					Assert.That(res[1].CreationDate, Is.EqualTo(DateTime.Parse("09 Apr 2019 14:30:35 GMT")));
+				}
+			}
+		}
+
 	}
 }
